move hdi storage account code
