+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-hdinsight/src/main/kotlin/com/microsoft/azure/hdinsight/spark/ui/SparkSubmissionJobUploadStorageAzureBlobCard.kt
@@ -45,8 +45,8 @@ import com.microsoft.intellij.forms.dsl.panel
 import com.microsoft.intellij.rxjava.IdeaSchedulers
 import com.microsoft.azure.toolkit.intellij.common.component.UIUtils
 import com.microsoft.intellij.ui.util.findFirst
-import com.microsoft.tooling.msservices.helpers.azure.sdk.StorageClientSDKManager
-import com.microsoft.tooling.msservices.model.storage.ClientStorageAccount
+import com.microsoft.azure.hdinsight.sdk.storage.StorageClientSDKManager
+import com.microsoft.azure.hdinsight.sdk.storage.model.ClientStorageAccount
 import org.apache.commons.lang3.StringUtils
 import org.apache.commons.lang3.StringUtils.EMPTY
 import org.apache.commons.lang3.exception.ExceptionUtils
