+++ Utils/hdinsight-node-common/src/com/microsoft/azure/hdinsight/serverexplore/AddHDInsightAdditionalClusterImpl.java
@@ -8,13 +8,12 @@ package com.microsoft.azure.hdinsight.serverexplore;
 import com.microsoft.azure.hdinsight.common.ClusterManagerEx;
 import com.microsoft.azure.hdinsight.common.StreamUtil;
 import com.microsoft.azure.hdinsight.sdk.common.HDIException;
-import com.microsoft.azure.hdinsight.sdk.storage.HDStorageAccount;
+import com.microsoft.azure.hdinsight.sdk.storage.implementation.HDStorageAccount;
 import com.microsoft.azuretools.azurecommons.helpers.AzureCmdException;
 import com.microsoft.azuretools.azurecommons.helpers.NotNull;
 import com.microsoft.azuretools.azurecommons.helpers.StringHelper;
-import com.microsoft.tooling.msservices.helpers.azure.sdk.StorageClientSDKManager;
-import com.microsoft.tooling.msservices.model.storage.ClientStorageAccount;
-import org.apache.commons.io.IOUtils;
+import com.microsoft.azure.hdinsight.sdk.storage.StorageClientSDKManager;
+import com.microsoft.azure.hdinsight.sdk.storage.model.ClientStorageAccount;
 import org.apache.http.auth.AuthScope;
 import org.apache.http.auth.UsernamePasswordCredentials;
 import org.apache.http.client.CredentialsProvider;
